Fixed doc examnples for nav kit react
