+++ playbook/app/pb_kits/playbook/pb_nav/docs/example.yml
@@ -7,7 +7,7 @@ examples:
   # - borderless_nav: No Borders
   # - subtle_nav: Subtle Variant
   - subtle_with_icons_nav: Subtle With Icons
-  - collapsible_nav: Subtle With Collapsible
+  - collapsible_nav: Collapsible Nav
   # - subtle_no_highlight_nav: Subtle No Highlight
   # - bold_vertical_nav: Bold Variant
   # - horizontal_nav: Horizontal Nav
@@ -25,8 +25,9 @@ examples:
   # - borderless_nav: No Borders
   # - subtle_nav: Subtle Variant
   - subtle_with_icons_nav: Subtle With Icons
-  - collapsible_nav: Subtle With Collapsible
-  - collapsible_nav_custom_icons: Subtle With Collapsible (No custom toggling)
+  - collapsible_nav: Collapsible Nav
+  - collapsible_nav_emphasize: Collapsible Nav with Emphasize Styling
+  - collapsible_nav_custom: Collapsible Nav With Custom Toggling
   # - subtle_no_highlight_nav: Subtle No Highlight
   # - bold_vertical_nav: Bold Variant
   # - horizontal_nav: Horizontal Nav
